requirements.txt
Flask
opencv-python
numpy
tensorflow
h5py

python-dotenv

